# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Studio Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentBreakdownItem } from './paymentBreakdownItem';
import type { StatusBreakdownItem } from './statusBreakdownItem';

export interface DashboardStats {
  totalOrdersToday: number;
  totalRevenueToday: number;
  pendingPickups: number;
  ordersInProgress: number;
  statusBreakdown: StatusBreakdownItem[];
  paymentBreakdown: PaymentBreakdownItem[];
}