# lib/api-zod/src/generated/types/hourlyItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Studio Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface HourlyItem {
  hour: number;
  count: number;
}