# lib/api-zod/src/generated/types/orderStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Studio Management System API
 * OpenAPI spec version: 0.1.0
 */

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  NEW: 'NEW',
  WAITING_PHOTOGRAPHY: 'WAITING_PHOTOGRAPHY',
  IN_PHOTOGRAPHY: 'IN_PHOTOGRAPHY',
  WAITING_EDITING: 'WAITING_EDITING',
  EDITING: 'EDITING',
  WAITING_PRINT: 'WAITING_PRINT',
  PRINTING: 'PRINTING',
  READY_FOR_DELIVERY: 'READY_FOR_DELIVERY',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED',
} as const;