# lib/api-zod/src/generated/types/orderUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Studio Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderService } from './orderService';
import type { PaymentMethod } from './paymentMethod';

export interface OrderUpdate {
  customerName?: string;
  customerMobile?: string;
  services?: OrderService[];
  paidAmount?: number;
  paymentMethod?: PaymentMethod;
  expectedDeliveryTime?: string;
  notes?: string;
}