# lib/api-zod/src/generated/types/paymentMethod.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Studio Management System API
 * OpenAPI spec version: 0.1.0
 */

export type PaymentMethod = typeof PaymentMethod[keyof typeof PaymentMethod];


export const PaymentMethod = {
  cash: 'cash',
  visa: 'visa',
  instapay: 'instapay',
  vodafone_cash: 'vodafone_cash',
} as const;